requirement.txt
amqp==2.5.2
asgiref==3.2.3
Babel==2.8.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.3
djangorestframework==3.11.0
flower==0.9.3
pycparser==2.19
python-decouple==3.3
python-jenkins==1.6.0
pytz==2019.3
rabbitmq==0.2.0
tornado==5.1.1
urllib3==1.25.8
vine==1.3.0
zipp==3.0.0
